# Change handling of orientations in geometries