… doc/source/guide/figures/parallel2d_geom.svg
… doc/source/guide/geometry_guide.rst
… doc/source/guide/guide.rst
… examples/solvers/chambolle_pock_tomography.py
… examples/solvers/chambolle_pock_tomography_tgv.py
… examples/solvers/conjugate_gradient_tomography.py
… examples/solvers/douglas_rachford_pd_tomography_tv.py
… examples/solvers/lbfgs_tomography.py
… examples/solvers/lbfgs_tomography_tv.py
… examples/solvers/nuclear_norm_tomography.py
… examples/solvers/proximal_gradient_wavelet_tomography.py
… examples/solvers/proximal_lang_tomography.py
… examples/tomo/README.md
… examples/tomo/anisotropic_voxels.py
… examples/tomo/backends/README.md
… examples/tomo/backends/astra_performance_cpu_parallel_2d_cg.py
… examples/tomo/backends/astra_performance_cuda_cone_3d_cg.py
… examples/tomo/backends/astra_performance_cuda_parallel_2d_cg.py
… examples/tomo/check_axes_parallel3d.py
… examples/tomo/checks/README.md
… examples/tomo/checks/check_axes_cone2d_bp.py
… examples/tomo/checks/check_axes_cone2d_fp.py
… examples/tomo/checks/check_axes_cone3d_bp.py
… examples/tomo/checks/check_axes_cone3d_fp.py
… examples/tomo/checks/check_axes_parallel2d_bp.py
… examples/tomo/checks/check_axes_parallel2d_fp.py
… examples/tomo/checks/check_axes_parallel3d_bp.py
… examples/tomo/checks/check_axes_parallel3d_fp.py
… examples/tomo/filtered_backprojection_cone_2d.py
… examples/tomo/filtered_backprojection_cone_2d_partial_scan.py