This preliminary PR fixes the issue with rotations as extensively discussed in #784. Tests currently fail since I didn't change the expected outcomes yet - before doing that I want to discuss the change.

The PR changes the following things:

- Detector initial position is now ~~`(-1, 0, 0)`~~ `(0, 1, 0)` by default, not `(1, 0, 0)`. The old value would mean that the detector would be illuminated from "behind". I think this default makes more sense.
- If a non-default initial detector position (for non-axial geometries) or symmetry axis (for axial geometries) is given, all vectors are rotated with the same matrix that rotates the default axis or detector position to the chosen one. This preserves the handedness of the coordinate system. Of course, everything can be overridden.
- ~~I added an `extra_rot` (maybe I should call it `extra_rotation`) parameter that performs a post-rotation of all vectors after the above initial rotation.~~ Initialization with a matrix that transforms the default configuration is possible with the alternative `frommatrix` constructor. This way the user has full control over the way things rotate.
- The geometries now have doctests that mostly explain the transformation of the coordinate systems. ~~Perhaps one could add tests for some other properties.~~ Done.
- The `__repr__` is updated and now looks much more readable.
- I documented the geometry concept of ODL and the geometric conventions. Perhaps the examples could be more explicit.

**Questions:**
- Are the changes OK?
- ~~The code for `__init__` and `__repr__` contains some remarkable overlap, in particular the determination of the initial rotation. Basically code [like this](https://github.com/kohr-h/odl/blob/95683e1dbf0665c079b8bd2637e9df8601f666b2/odl/tomo/geometry/parallel.py#L464-L489). It could be factored out, but I don't quite know where to put it. It's almost the same in 2 and 3 dimensions, so we could put the code in `Geometry`, but then I don't know if all geometries want to handle orientations like that.
   So the question is: should we factor out and where should we put the code?
   One way to do this is to add a private method to `Geometry` that does this initialization step, and then let subclasses decide whether or not they want to use it. No code should then assume that the method was called.~~
  This whole issue will go away when `matrix` excludes the other stuff.
- Anything missing in the geometry documentation?

**TODOs:**
- [x] Fix #772 here by incorporating #773
- [x] Run examples to check the impact
- [x] Change detector default position to `(0, 1, 0)`. This gives `(x, z)` as natural initial coordinate system on the detector while being visually and numerically intuitive (no weird minus sign etc.)
- [x] Fix documentation
- [x] Make matrix argument for geometries exclude `det_axes_init` and the likes to avoid confusion and overly complicated code
- [x] Rename `extra_rot` to `matrix` that indicates a broader use 
- [x] Add translation
- [x] Fix misc review comments
- [x] Avoid `rot90` in 2d projections / backprojections
- [x] Write README in `examples/tomo/checks`
- [x] Update `indices=[...]` syntax for `show` in examples
- [x] Add `cone_beam_geometry` helper similar to `parallel_beam_geometry`
- [x] Remove unnecessary `asarray()` calls
- [x] Regroup examples list in README