@adler-j 2017-04-10
Short pointer: This should incorporate the work in #773 so that we get proper tests once and for all.

I'll try to get a review done soon and answer the questions.