@kohr-h 2017-04-10
> Short pointer: This should incorporate the work in #773 so that we get proper tests once and for all.

Yes, thx. That should go in here. What's the status of that one? Tests added that fail due to a bug?