@adler-j 2017-04-10
Yes basically i found a bug and added a test that (correctly) fails due to this bug. Never got around to actually fixing it in the geometries.