@kohr-h 2017-04-25
> So this is my next major review target. Is it ready for a review?

Yes, apart from the trivial rebase conflicts and the failing tests (since I didn't adapt them yet).

It's actually not that big of a PR, most of the stuff are doctests.