@kohr-h 2017-05-16
So I fixed and extended the tests a bit, should all work now. My own comments are still valid, need to fix those tomorrow. Apart from that, I'm quite happy with the shape of the branch. The remaining TODOs are on the top for reference.