@kohr-h 2017-05-18
> The "extra rot" should, imo, happen inside rotation_matrix, that way we don't need to keep 3 coordinate systems in mind all the time, but can stay at two.

The problem with that would be that `det_pos_init != det_point_posiiton(0, ...)` which would be super-counter-intuitive. I'm not sure if there's a zero-surprise option. At least currently the doc says that the matrix is applied *afterwards*.
An alternative would be to allow the matrix variant only if the other arguments are default, in a sense as a second signature. That would also simplify `__repr__`.