@adler-j 2017-05-19
Well it is called "extra" for a reason, no? But i sure agree with your point, it is confusing. Frankly the alternative of not allowing a mixture is perhaps more simple.