@kohr-h 2017-05-23
I've updated this stuff, the main change being the (now passing) test in #773 and a bunch of example scripts that can be run for visual comparison with plain summing along axes. That's not the same as a test, but whenever there's a change that preserves the convention, all the examples should still look correct.