@kohr-h 2017-05-25
Everything fixed now. Major changes since review:

- more and simpler doctests added,
- a global `translation` can now be provided,
- an `init_matrix` can be specified for geometric transformation, but not together with the other arguments,
- 3d default is now `(0, 1, 0)` for detector position and `(1, 0, 0), (0, 0, 1)` for the detector axes,
- documentation updated, with some figures,
- repr simplified greatly by storing the initial arguments (no more reconstruction of arguments)
- more comprehensive unit tests, should be pretty good now,
- new example scripts that can be run to visually check if the axes are correct,
- some more small things

I'd like to do one last thing, namely remove the `fanbeam` module and add the stuff to `conebeam`. IMO there's no good reason to treat these geometries different from the parallel ones, where 2d and 3d is in the same module.

I guess a new review is necessary...