@kohr-h 2017-05-26
> Should I review again?

Would be nice :-)