@kohr-h 2017-05-26
Parallel 2d geometries seem to be broken when the volume is not centered around zero. No big surprise because that info is never communicated to ASTRA. Did that ever work? I guess that requires an exception when somebody tries this.