@adler-j 2017-05-26
bug report:

https://github.com/odlgroup/odl/issues/359

fix in wait:

https://github.com/astra-toolbox/astra-toolbox/pull/69