@kohr-h 2017-06-12
Thanks for the review. I'll try to push this asap since I feel it's important to have the geometries consistent and properly documented regarding conventions etc.