@kohr-h 2017-06-16
> I must admit I did NOT re-read all of it but these are some main comments.
> 
> Is there anything in particular you need re-reviewed?

Not now, it's way too early. I first need to port all classes to the alternative constructor pattern and fix the repr methods, the doctests and then finally the unit tests.
But you seem to be happy with this change, so for now the main focus can be the math.