@kohr-h 2017-06-19
All fixed, in particular avoiding rotation of data in 2d now. Ready for final assessment.