@adler-j 2017-06-20
517d62e should close https://github.com/odlgroup/odl/issues/903.

will review.