@kohr-h 2017-06-27
Merged a PR that adds vectorization of the geometry methods, something that's deeply needed. It cuts down the time to convert a geometry with 1000 angles to ASTRA vec from over 200 ms (!) to about 1 ms.
That brings with it a couple of new TODOs which I added at the top.