@adler-j 2017-06-27
The geometries already cached these so that wasn't so bad, but a nice change regardless, if nothing else for usability. I'm doing a review of this atm but it looks mostly good.