@kohr-h 2017-06-27
> The geometries already cached these so that wasn't so bad, but a nice change regardless, if nothing else for usability. I'm doing a review of this atm but it looks mostly good.

Well we are doing single-slice reconstruction which sets up a new geometry for each requested slice, so those geometries are only used once. In that case, caching doesn't help and 200 ms are an actual bottleneck. Otherwise: true.