@adler-j 2017-06-28
> Regarding the vectorization, I had to rush that a bit because I needed it. Since it's still WIP I'll probably rebase it out again and make a separate PR for that. Agreed?

Fine by me