@kohr-h 2017-07-02
Fixed all comments, added a test for `cone_beam_geometry` and rebased, I feel this should go in asap.

A remark on 98e43fe:
In ASTRA 3D geometries I now switch the `u` and `v` detector coordinates. This has the advantage that `v` ends up being the fastest-changing axis in the projection data returned by ASTRA, which is the right place. So instead of changing all axes and making the data completely discontiguous, we only have to swap the first two, resulting in much faster `ascontiguousarray` runtime.

Compare this result from running [`line_profiler`](https://github.com/rkern/line_profiler) on the [filtered_backprojection_cone_3d_partial_scan.py](https://github.com/odlgroup/odl/blob/master/examples/tomo/filtered_backprojection_cone_3d_partial_scan.py) example.

### Before the change:
```
Total time: 1.42126 s
File: /export/scratch1/kohr/git/odl/odl/tomo/backends/astra_cuda.py
Function: call_forward at line 65

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    65                                               @profile
    66                                               def call_forward(self, vol_data, out=None):
    67                                                   """Run an ASTRA forward projection on the given data using the GPU.
    68
    69                                                   Parameters
    70                                                   ----------
    71                                                   vol_data : `reco_space` element
    72                                                       Volume data to which the projector is applied.
    73                                                   out : `proj_space` element, optional
    74                                                       Element of the projection space to which the result is written. If
    75                                                       ``None``, an element in `proj_space` is created.
    76
    77                                                   Returns
    78                                                   -------
    79                                                   out : ``proj_space`` element
    80                                                       Projection data resulting from the application of the projector.
    81                                                       If ``out`` was provided, the returned object is a reference to it.
    82                                                   """
    83         1            8      8.0      0.0          assert vol_data in self.reco_space
    84         1            0      0.0      0.0          if out is not None:
    85                                                       assert out in self.proj_space
    86                                                   else:
    87         1           44     44.0      0.0              out = self.proj_space.element()
    88
    89                                                   # Copy data to GPU memory
    90         1            2      2.0      0.0          if self.geometry.ndim == 2:
    91                                                       astra.data2d.store(self.vol_id, vol_data.asarray())
    92         1            1      1.0      0.0          elif self.geometry.ndim == 3:
    93         1        31932  31932.0      2.2              astra.data3d.store(self.vol_id, vol_data.asarray())
    94                                                   else:
    95                                                       raise RuntimeError('unknown ndim')
    96
    97                                                   # Run algorithm
    98         1       888968 888968.0     62.5          astra.algorithm.run(self.algo_id)
    99
   100                                                   # Copy result to host
   101         1            9      9.0      0.0          if self.geometry.ndim == 2:
   102                                                       out[:] = self.out_array
   103         1            1      1.0      0.0          elif self.geometry.ndim == 3:
   104         1           24     24.0      0.0              out[:] = np.rollaxis(self.out_array, 0, 3).reshape(
   105         1       500268 500268.0     35.2                  self.proj_space.shape)
   106
   107                                                   # Fix scaling to weight by pixel size
   108         1            5      5.0      0.0          if isinstance(self.geometry, Parallel2dGeometry):
   109                                                       # parallel2d scales with pixel stride
   110                                                       out *= 1 / float(self.geometry.det_partition.cell_volume)
   111
   112         1            1      1.0      0.0          return out



Total time: 11.5246 s
File: /export/scratch1/kohr/git/odl/odl/tomo/backends/astra_cuda.py
Function: call_backward at line 211

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   211                                               @profile
   212                                               def call_backward(self, proj_data, out=None):
   213                                                   """Run an ASTRA back-projection on the given data using the GPU.
   214
   215                                                   Parameters
   216                                                   ----------
   217                                                   proj_data : `proj_space` element
   218                                                       Projection data to which the back-projector is applied.
   219                                                   out : `reco_space` element, optional
   220                                                       Element of the reconstruction space to which the result is written.
   221                                                       If ``None``, an element in ``reco_space`` is created.
   222
   223                                                   Returns
   224                                                   -------
   225                                                   out : ``reco_space`` element
   226                                                       Reconstruction data resulting from the application of the
   227                                                       back-projector. If ``out`` was provided, the returned object is a
   228                                                       reference to it.
   229                                                   """
   230         2           11      5.5      0.0          assert proj_data in self.proj_space
   231         2            2      1.0      0.0          if out is not None:
   232                                                       assert out in self.reco_space
   233                                                   else:
   234         2           96     48.0      0.0              out = self.reco_space.element()
   235
   236                                                   # Copy data to GPU memory
   237         2            3      1.5      0.0          if self.geometry.ndim == 2:
   238                                                       astra.data2d.store(self.sino_id, proj_data.asarray())
   239         2            2      1.0      0.0          elif self.geometry.ndim == 3:
   240         2           13      6.5      0.0              if self.geometry.motion_params.ndim > 1:
   241                                                           shape = (-1,) + self.geometry.det_partition.shape
   242                                                           reshaped_proj_data = proj_data.asarray().reshape(shape)
   243                                                           swapped_proj_data = np.ascontiguousarray(
   244                                                               np.rollaxis(reshaped_proj_data, 2, 0))
   245                                                       else:
   246         2            2      1.0      0.0                  swapped_proj_data = np.ascontiguousarray(
   247         2     10804843 5402421.5     93.8                      np.rollaxis(proj_data.asarray(), 2, 0))
   248
   249         2       139637  69818.5      1.2              astra.data3d.store(self.sino_id, swapped_proj_data)
   250
   251                                                   # Run algorithm
   252         2       504619 252309.5      4.4          astra.algorithm.run(self.algo_id)
   253
   254                                                   # Copy result to CPU memory
   255         2        59080  29540.0      0.5          out[:] = self.out_array
   256
   257                                                   # Fix scaling to weight by pixel/voxel size
   258         2            4      2.0      0.0          out *= astra_cuda_bp_scaling_factor(
   259         2        16283   8141.5      0.1              self.proj_space, self.reco_space, self.geometry)
   260
   261         2            3      1.5      0.0          return out
```

### After the change:
```
Total time: 1.06737 s
File: /export/scratch1/kohr/git/odl/odl/tomo/backends/astra_cuda.py
Function: call_forward at line 65

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    65                                               @profile
    66                                               def call_forward(self, vol_data, out=None):
    67                                                   """Run an ASTRA forward projection on the given data using the GPU.
    68
    69                                                   Parameters
    70                                                   ----------
    71                                                   vol_data : `reco_space` element
    72                                                       Volume data to which the projector is applied.
    73                                                   out : `proj_space` element, optional
    74                                                       Element of the projection space to which the result is written. If
    75                                                       ``None``, an element in `proj_space` is created.
    76
    77                                                   Returns
    78                                                   -------
    79                                                   out : ``proj_space`` element
    80                                                       Projection data resulting from the application of the projector.
    81                                                       If ``out`` was provided, the returned object is a reference to it.
    82                                                   """
    83         1            8      8.0      0.0          assert vol_data in self.reco_space
    84         1            1      1.0      0.0          if out is not None:
    85                                                       assert out in self.proj_space
    86                                                   else:
    87         1           47     47.0      0.0              out = self.proj_space.element()
    88
    89                                                   # Copy data to GPU memory
    90         1            2      2.0      0.0          if self.geometry.ndim == 2:
    91                                                       astra.data2d.store(self.vol_id, vol_data.asarray())
    92         1            1      1.0      0.0          elif self.geometry.ndim == 3:
    93         1        31894  31894.0      3.0              astra.data3d.store(self.vol_id, vol_data.asarray())
    94                                                   else:
    95                                                       raise RuntimeError('unknown ndim')
    96
    97                                                   # Run algorithm
    98         1       818952 818952.0     76.7          astra.algorithm.run(self.algo_id)
    99
   100                                                   # Copy result to host
   101         1            8      8.0      0.0          if self.geometry.ndim == 2:
   102                                                       out[:] = self.out_array
   103         1            1      1.0      0.0          elif self.geometry.ndim == 3:
   104         1           15     15.0      0.0              out[:] = np.swapaxes(self.out_array, 0, 1).reshape(
   105         1       216429 216429.0     20.3                  self.proj_space.shape)
   106
   107                                                   # Fix scaling to weight by pixel size
   108         1            7      7.0      0.0          if isinstance(self.geometry, Parallel2dGeometry):
   109                                                       # parallel2d scales with pixel stride
   110                                                       out *= 1 / float(self.geometry.det_partition.cell_volume)
   111
   112         1            1      1.0      0.0          return out

Total time: 0.772498 s
File: /export/scratch1/kohr/git/odl/odl/tomo/backends/astra_cuda.py
Function: call_backward at line 209

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   209                                               @profile
   210                                               def call_backward(self, proj_data, out=None):
   211                                                   """Run an ASTRA back-projection on the given data using the GPU.
   212
   213                                                   Parameters
   214                                                   ----------
   215                                                   proj_data : `proj_space` element
   216                                                       Projection data to which the back-projector is applied.
   217                                                   out : `reco_space` element, optional
   218                                                       Element of the reconstruction space to which the result is written.
   219                                                       If ``None``, an element in ``reco_space`` is created.
   220
   221                                                   Returns
   222                                                   -------
   223                                                   out : ``reco_space`` element
   224                                                       Reconstruction data resulting from the application of the
   225                                                       back-projector. If ``out`` was provided, the returned object is a
   226                                                       reference to it.
   227                                                   """
   228         2           16      8.0      0.0          assert proj_data in self.proj_space
   229         2            1      0.5      0.0          if out is not None:
   230                                                       assert out in self.reco_space
   231                                                   else:
   232         2          114     57.0      0.0              out = self.reco_space.element()
   233
   234                                                   # Copy data to GPU memory
   235         2            4      2.0      0.0          if self.geometry.ndim == 2:
   236                                                       astra.data2d.store(self.sino_id, proj_data.asarray())
   237         2            2      1.0      0.0          elif self.geometry.ndim == 3:
   238         2           34     17.0      0.0              shape = (-1,) + self.geometry.det_partition.shape
   239         2           48     24.0      0.0              reshaped_proj_data = proj_data.asarray().reshape(shape)
   240         2            2      1.0      0.0              swapped_proj_data = np.ascontiguousarray(
   241         2       208983 104491.5     27.1                  np.swapaxes(reshaped_proj_data, 0, 1))
   242         2       140599  70299.5     18.2              astra.data3d.store(self.sino_id, swapped_proj_data)
   243
   244                                                   # Run algorithm
   245         2       347233 173616.5     44.9          astra.algorithm.run(self.algo_id)
   246
   247                                                   # Copy result to CPU memory
   248         2        58990  29495.0      7.6          out[:] = self.out_array
   249
   250                                                   # Fix scaling to weight by pixel/voxel size
   251         2            4      2.0      0.0          out *= astra_cuda_bp_scaling_factor(
   252         2        16465   8232.5      2.1              self.proj_space, self.reco_space, self.geometry)
   253
   254         2            3      1.5      0.0          return out
```

For forward projection, we gain a little bit:
- 200 ms instead of 500 for the storage of the swapped data
- Accounts for 20 % of the run time instead of 35 %

But in the backprojection the difference is huge:
- 100 ms instead of **5400 ms** for making the data contiguous (the explicit `ascontiguousarray` is not responsible)
- 27 % run time instead of **94 %**
- Total time for BP 380 ms instead of **5750 ms**.

Perhaps this is an issue with `ascontiguousarray` since assigning directly from a discontiguous array to a contiguous one doesn't take that long (as in the FP).